pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "newsletter-automation"
version = "1.0.0"
description = "Automate newsletter processing with AI synthesis, Notion, and Gmail"
readme = "README_GENERIC.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Newsletter Automation Contributors"}
]
keywords = ["newsletter", "automation", "gmail", "notion", "perplexity", "ai"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Communications :: Email",
    "Topic :: Office/Business",
]

dependencies = [
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.1.1",
    "google-api-python-client>=2.80.0",
    "notion-client>=2.0.0",
    "requests>=2.28.0",
    "html2text>=2020.1.16",
]

[project.urls]
Homepage = "https://github.com/votreusername/newsletter-automation"
Documentation = "https://github.com/votreusername/newsletter-automation/blob/main/README_GENERIC.md"
Repository = "https://github.com/votreusername/newsletter-automation.git"
Issues = "https://github.com/votreusername/newsletter-automation/issues"

[project.scripts]
newsletter-automation = "src.newsletter_automation.newsletter_automation:main"

[tool.setuptools]
packages = ["src", "src.newsletter_automation"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 100
